modular_code/mementoai/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sentence-transformers==4.1.0
pinecone
google-generativeai==0.8.5
celery==5.5.2
redis==6.0.0
streamlit==1.35.0
validators==0.35.0
python-dotenv==1.1.0
requests==2.32.3
gitpython==3.1.44
radon==6.0.1
lizard==1.17.10
plotly==5.18.0
